refactor: react-query 사용 훅 이름 통일 및 구조 변경
