--- src/features/schedules/api/getSchedules.ts
@@ -1,14 +1,0 @@
-import { useQuery } from "react-query";
-
-import type { ScheduleSimpleResponse } from "../types";
-
-import { axios } from "@/lib/axios";
-
-export const getSchedules = (): Promise<ScheduleSimpleResponse[]> => {
-  return axios.get(`/schedules`);
-};
-
-export const useSchedulesData = () => {
-  const { data = [] } = useQuery(["schedules"], getSchedules);
-  return { data };
-};
