--- src/features/schedules/api/modifySchedule.ts
@@ -1,30 +1,0 @@
-import { useMutation, useQueryClient } from "react-query";
-import { useNavigate } from "react-router-dom";
-
-import type { ScheduleModificationRequest } from "../types";
-
-import { axios } from "@/lib/axios";
-
-export type ModifyScheduleDTO = {
-  data: ScheduleModificationRequest;
-  scheduleId: number;
-};
-
-export const modifySchedule = ({
-  data,
-  scheduleId,
-}: ModifyScheduleDTO): Promise<number> => {
-  return axios.put(`/schedules/${scheduleId}`, data);
-};
-
-export const useModifySchedule = () => {
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-
-  return useMutation(modifySchedule, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(["schedules"]);
-      navigate(`/schedules`);
-    },
-  });
-};
